Выносит функции и объекты фильтрации в утилиты
