+++ src/const.js
@@ -1,0 +1,8 @@
+const FilterType = {
+  ALL_MOVIES: `All movies`,
+  WATCHLIST: `Watchlist`,
+  HISTORY: `History`,
+  FAVORITES: `Favorites`,
+};
+
+export {FilterType};
